Max distance should be float too

+++ src/config.h
@@ -24,7 +24,7 @@ namespace config {
   // Output config SDL, Mustein, UART
 
   // Sphere Tracer limits
-  constexpr int   traceMaxDistance  = 80.0f;
+  constexpr float traceMaxDistance  = 80.0f;
   constexpr int   traceMaxSteps     = 80;
   constexpr float minObjectDistance = 0.01f;
 }
